Refactor Tron module and make gasUsed and gas int instead of Amount
