+++ lib/src/crypto/tron/tron.dart
@@ -2,5 +2,4 @@ library tron;
 
 export 'repositories/tron_http_repository.dart';
 export 'repositories/tronscan_repository.dart';
-export 'entities/tron_transaction.dart';
 export 'entities/tron_address.dart';
