+++ test/ci/tron/parameter_test.dart
@@ -1,6 +1,6 @@
 import 'package:test/test.dart';
+import 'package:walletkit_dart/src/crypto/tron/entities/tron_contract_type.dart';
 import 'package:walletkit_dart/src/crypto/tron/entities/tron_parameter_type.dart';
-import 'package:walletkit_dart/src/crypto/tron/entities/tron_transaction.dart';
 import 'package:walletkit_dart/src/domain/entities/amount.dart';
 import 'package:walletkit_dart/src/domain/extensions.dart';
